test: run common set of tests for various setups
